Whitespace and stan/cs fixes.

+++ src/Utility/Util.php
@@ -20,10 +20,9 @@ class Util {
 		return $randomString;
 	}
 
-	public static function getPercentButAvoid100UntillComplete(int $value, int $max): int
-	{
+	public static function getPercentButAvoid100UntillComplete(int $value, int $max): int {
 		assert($value <= $max);
-		$result = round($value / $max);
+		$result = (int) round($value / $max);
 		if ($result == 100 && $value < $max) {
 			return 99;
 		}
